+plural Read & Update tasks

+++ src/lib/components/AppPulloutMenu.svelte
@@ -1,5 +1,8 @@
 <script lang="ts">
 	import { dev as DEVELOPMENT } from '$app/environment';
+	import { invalidate } from '$app/navigation';
+	import { page } from '$app/state';
+	import { getUser } from '$lib/API/Auth';
 	import supabase from '$lib/API/SupabaseClient';
 	import BugReportModal from './BugReportModal.svelte';
 	import PulloutBlock from './overlays/Pullout.svelte';
@@ -7,8 +10,11 @@
 	let showPullout = $state(false);
 
 	async function deleteAllTasks() {
-		await supabase.from('tasks').delete();
+		const user = await getUser();
+		if(!user) throw new Error(`No user could be found to delete all tasks...`);
+		await supabase.from('tasks').delete().eq('user_id', user.id); // basically WHERE true
 		showPullout = false;
+		window.location.reload();
 	}
 </script>
